requirements.txt
pymupdf
openai
pandas
tqdm
python-dotenv
